OMAP4: USB: Fixed tshut reboot issue
Wakeup enable was set for usbb1_ulpitll_dir.usbb1_ulpiphy_dir. This is also
muxed with gpio 86. When a wakeup event was detected as part of the
gpio runtime resume, it set the gpio 86 event causing a tshut irq event.
We now rely on the wakeup from pin usbb1_ulpitll_dat0.usbb1_ulpiphy_dat0 for
USB wakeup events.

Change-Id: Ic32a976118944e45b91c0bdce6099ab090da3ae7
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>