OMAP2+: DVFS: optimize DVFS scale
Currently, the s/w model assumes that when _dvfs_scale is called
we definitely need to do a scale. However when IVA/MPU scales, due to
dependency, _dvfs_scale is invoked for core domain. This unfortunately
does result in SR being disabled, re-enabled, an attempt of voltagescale
and a scan of frequency changes to take place, when there is NO need
for any of these to take place.

we can optimize by doing a pre-scan of the data structures prior to
doing the timeconsuming transition operations

Change-Id: I76e1989acc077092bd7af3027c6fe05adaaffca1
Signed-off-by: Nishanth Menon <nm@ti.com>