OMAP3+: VP: do not allow voltage lower than retention voltage
Do not allow the Voltage Processor to go below the retention
voltage even if HLOS or SmartReflex AVS module requests it.

This is to prevent unsafe voltages from ever being achieved
on the device. if the voltage domain cannot function below
the retention voltage during inactive retention power state,
there is hardly a possibility for the voltagedomain to function
at lower voltages in active scenarios

Change-Id: Idd1b199431c15002cf212d84e26e83d1940f9387
Signed-off-by: Nishanth Menon <nm@ti.com>