omap: rpmsg: switch over vring memory to use dynamic pool for dsp
The carveout memory for DSP is being switched over to use the
dynamically reserved memory pool. The memory addresses used to
configure the vring memory is therefore being switched over to
use the dynamic memory pool.

The vring memory is allocated in the rpmsg layer, while the
memory for other firmware sections is assigned by the remoteproc
driver. The remoteproc driver adjusts for the utilization of
this memory automatically through the presence of the vring
carveout section as the first of the carveout sections in the
resource table.

Change-Id: I5156028450fc7aec813bf3ce9b328acf06059de1
Signed-off-by: Jesse Villarreal <jesse.villarreal@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>

Conflicts:

	arch/arm/plat-omap/omap_rpmsg.c