This contains two commits: Both of which I kanged from Morific's git.
omap4: Add basic earlysuspend hotplug driver

This driver hotplugs CPU1 offline when early_suspend is called
and onlines it when late_resume is called.
This is advantageous to power saving as CPU0 can then enter
OSWR without agreeing with CPU1, which saves power in certain
scenarios such as screen off audio playback and also when the
device is woken by a partial wakelock to perform a routine task.

omap: earlysuspend: Use delayed work for hotplugging
Use deferrable delayed work to hotplug CPU1 offline and online.
Use a delay of HZ (1 second) for both offlining and onlining.
This may fix the lockups some devices are experiencing.
Also remove the locking as we are using schedule_delayed_work_on.

Signed-off-by: Will Tisdale <willtisdale@gmail.com>